requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
Django==6.0
django-filter==25.2
django-htmx==1.27.0
djangorestframework==3.16.1
Flask==3.1.2
flatbuffers==25.9.23
gunicorn==23.0.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
itsdangerous==2.2.0
JayDeBeApi==1.2.3
Jinja2==3.1.6
jpype1==1.6.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.45.1
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6
numba==0.62.1
numpy==2.2.1
onnxruntime==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
protobuf==6.33.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyMatting==1.1.14
pyodbc==5.3.0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rembg==2.0.68
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.25.2
scipy==1.13.1
setuptools==80.9.0
sqlparse==0.5.4
sympy==1.14.0
tifffile==2025.10.16
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata
urllib3==2.5.0
Werkzeug==3.1.4
wheel==0.45.1
whitenoise==6.11.0